requirements.txt
click==8.1.3
matplotlib==3.6.2
numpy==1.23.5
python-dotenv==0.21.0
scikit_learn==1.2.0
setuptools==65.6.3
torch>=1.13.0+cu116
torchvision>=0.14.0+cu116
